dataloader: Fix type of nsamples in create_sparse_batch_stream
The nsamples parameter was an int which caused an overflow when using
3B samples.

+++ dataloader/dataloader.c
@@ -173,7 +173,7 @@ static void add_sample_to_batch(bool use_factorizer, int sample_idx,
 }
 
 EXPORT struct stream* CDECL create_sparse_batch_stream(const char* filename,
-                                                       int nsamples,
+                                                       uint64_t nsamples,
                                                        int batch_size,
                                                        int use_factorizer)
 {
